+++ config/dev.exs
@@ -55,4 +55,4 @@ config :phoenix, :plug_init_mode, :runtime
 
 config :swoosh, :api_client, false
 
-config :wax_, origin: "http://localhost:4000", rp_id: :auto
+config :wax_, origin: "http://localhost:4000", rp_id: "localhost"